% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_model}
\alias{run_model}
\title{run models}
\usage{
run_model(
  model = SIRmodel,
  parms = c(mu = 0, beta = 2, gamma = 1),
  yini = c(S = 0.999, I = 0.001, R = 0),
  tvec = seq(0, 20, by = 0.1)
)
}
\description{
run models
}
